Update main.py
Se han realizado cambios en el script para mostrar la cantidad total de posts procesados en cada página. Además, ahora el script evita la descarga de imágenes que ya han sido guardadas localmente para evitar repeticiones.

+++ main.py
@@ -4,7 +4,7 @@ from bs4 import BeautifulSoup
 from urllib.parse import urljoin, urlparse, unquote
 
 # Configuración de la URL del sitio
-wordpress_url = 'https://webdestino.com/blog/'
+wordpress_url = 'https://thefinngroup.com.au/finn-blog/'
 
 # Ruta local para guardar las imágenes
 image_folder = os.path.join(os.getcwd(), 'export', 'images')
@@ -24,8 +24,13 @@ while True:
     html_content = response.content
     soup = BeautifulSoup(html_content, 'html.parser')
 
+    print(f'Procesando la página {page_number}...')
+
     # Encontrar enlaces a las páginas de cada post
     post_links = [a['href'] for a in soup.find_all('a', class_='link-to-post')]
+
+    # Contador de posts procesados en la página actual
+    posts_processed = 0
 
     # Iterar sobre los enlaces de los posts
     for post_link in post_links:
@@ -44,20 +49,22 @@ while True:
             if "/collect/" in featured_image_url:
                 print(f'Se omite la imagen destacada: {featured_image_url}')
             else:
-                # Descargar y guardar la imagen destacada
+                # Descargar y guardar la imagen destacada si no existe localmente
                 featured_image_filename = os.path.join(image_folder, os.path.basename(urlparse(featured_image_url).path))
-                os.makedirs(os.path.dirname(featured_image_filename), exist_ok=True)  # Crear carpetas si no existen
-                featured_image_data = requests.get(featured_image_url).content
-                with open(featured_image_filename, 'wb') as img_file:
-                    img_file.write(featured_image_data)
-                    print(f'Imagen destacada guardada: {featured_image_filename}')
+                if not os.path.exists(featured_image_filename):
+                    os.makedirs(os.path.dirname(featured_image_filename), exist_ok=True)  # Crear carpetas si no existen
+                    featured_image_data = requests.get(featured_image_url).content
+                    with open(featured_image_filename, 'wb') as img_file:
+                        img_file.write(featured_image_data)
+                        print(f'Imagen destacada guardada: {featured_image_filename}')
+                        posts_processed += 1
 
         # Obtener las imágenes internas
         internal_images = post_soup.find_all('img')
 
         for internal_image in internal_images:
             internal_image_url = internal_image['src']
-            # Verificar si la URL contiene "/collect/" antes de continuar ""OPCIONAL SOLO PARA IGNORAR ALGUNAS IMAGENES VISIBLES DENTRO DE LOS POST Y QUE NO CORRESPONDAN A LA ENTRADA""
+            # Verificar si la URL contiene "/collect/" antes de continuar
             if "/collect/" in internal_image_url:
                 print(f'Se omite la imagen interna: {internal_image_url}')
             else:
@@ -74,12 +81,16 @@ while True:
                 image_folder_path = os.path.join(image_folder, image_path)
                 os.makedirs(os.path.dirname(image_folder_path), exist_ok=True)  # Crear carpetas si no existen
 
-                # Descargar y guardar la imagen interna
+                # Descargar y guardar la imagen interna si no existe localmente
                 internal_image_filename = os.path.join(image_folder, image_path.replace('/', '\\'))
-                internal_image_data = requests.get(internal_image_url).content
-                with open(internal_image_filename, 'wb') as img_file:
-                    img_file.write(internal_image_data)
-                    print(f'Imagen interna guardada: {internal_image_filename}')
+                if not os.path.exists(internal_image_filename):
+                    internal_image_data = requests.get(internal_image_url).content
+                    with open(internal_image_filename, 'wb') as img_file:
+                        img_file.write(internal_image_data)
+                        print(f'Imagen interna guardada: {internal_image_filename}')
+                        posts_processed += 1
+
+    print(f'Página {page_number} procesada. Total de posts procesados: {posts_processed}')
 
     page_number += 1
